add ConEmu Launcher extension
ConEmu Launcher is similar to Console Launcher however it uses
a concept of a single task rather than multiple tabs. It also allows
to specify the literal command line switches to be passed to ConEmu.
The ideas of hiding the menu commands and validating the settings
upon apply were introduced while developing this extension.